Rename criterion.py to fit_criterion.py
