+++ baybe/surrogates/gaussian_process/components/fit_criterion.py
