+++ baybe/surrogates/gaussian_process/presets/edbo.py
@@ -21,7 +21,7 @@ from baybe.parameters.selectors import (
 from baybe.parameters.substance import SubstanceParameter
 from baybe.priors.basic import GammaPrior
 from baybe.searchspace.discrete import SubspaceDiscrete
-from baybe.surrogates.gaussian_process.components.criterion import (
+from baybe.surrogates.gaussian_process.components.fit_criterion import (
     FitCriterion,
     PlainFitCriterionFactory,
 )
