+++ baybe/surrogates/gaussian_process/presets/edbo_smoothed.py
@@ -18,7 +18,7 @@ from baybe.parameters.selectors import (
     to_parameter_selector,
 )
 from baybe.priors.basic import GammaPrior
-from baybe.surrogates.gaussian_process.components.criterion import (
+from baybe.surrogates.gaussian_process.components.fit_criterion import (
     FitCriterion,
     PlainFitCriterionFactory,
 )
